+++ src/App.tsx
@@ -1,16 +1,11 @@
+import {useState} from 'react';
 import Header from "./components/Header";
 import UserInput from "./components/UserInput";
 import Results from "./components/Results";
+import type { UserInputType } from "./types";
 import "./App.scss";
 
-function App() {
-
-  type UserInputType = {
-    initialInvestment: number;
-    annualInvestment: number;
-    expectedReturn: number;
-    duration: number;
-  };
+function App() {  
 
   const [userInput, setUserInput] = useState<UserInputType>({
     initialInvestment: 1000,
@@ -40,13 +35,13 @@ function App() {
       newInput[inputIdentifier] = newValue;
       return newInput;
     });
-  }
+  } // end of function handleChange
 
   return (
     <>
       <Header />
       <UserInput onChange={handleChange} userInput={userInput} />
-      <Results />
+      <Results input={userInput} />
     </>
   );
 }
