+++ src/components/Results.tsx
@@ -1,9 +1,16 @@
-function Results() {
+import type { UserInputType } from "../types";
+
+type Props = {
+  input: UserInputType;
+};
+
+function Results({ input }: Props) {
+  console.log(input);
   return (
     <>
       <p>Results...</p>
     </>
-  )
+  );
 }
 
 export default Results;
