+++ src/types.ts
@@ -1,0 +1,8 @@
+type UserInputType = {
+  initialInvestment: number;
+  annualInvestment: number;
+  expectedReturn: number;
+  duration: number;
+};
+
+export type {UserInputType};
